requirements.txt
fastapi
uvicorn
python-multipart
jinja2
torch
openai-whisper
demucs
pydub
aiofiles
soundfile